CLN: resolvers must be checked before the calling frame
Otherwise an expression like

df['index']

will blow the stack. This is because 'index' is a valid expression (if it isn't
a column name) so an neverending mutual recursion will occur between looking
up a value and trying to query the expression, because look up is done through
DataFrame.__getitem__().